6/26/17 { +   Added prayer bonus calculation to all combat calculators +   Added weight per item and weight calculation for energy draining while running. +   Added player resting via orbs 	-   As well as the getting up animation +   Added item weights to the stats popup in the equipment menu +   Added protection prayer effects and deflection prayer effects 	-   Different graphics for magic/ranged/melee

+++ src/org/redrune/game/node/entity/player/data/PlayerVariables.java
@@ -41,7 +41,7 @@ public final class PlayerVariables {
 	 */
 	@Getter
 	@Setter
-	private int runEnergy = 100;
+	private double runEnergy = 100;
 	
 	/**
 	 * If the run button has been toggled on
